fix: rename files with footnotes
