+++ src/footnote-plugin/encoding.js
