+++ src/footnote-plugin/footnote-view.js
