+++ src/footnote-plugin/utils.js
